refactor(app): Shifts to modular approach to keep seed and tests scripts out of src folder.
